inventory-service/requirements.txt
Flask==2.1.0
pymongo==3.12.0
redis==3.5.3
requests==2.26.0
python-dotenv==0.19.0
gunicorn==20.1.0

Flask==2.2.2                # The web framework
Flask-CORS==3.0.10          # CORS support for Flask
PyJWT==2.6.0                # JWT handling
pymongo==4.3.2              # MongoDB driver for Python
redis==4.5.5                # Redis client for Python
python-dotenv==1.0.0        # For loading environment variables from a .env file
